test cross building in our CI rule (#396)
* test cross building in our CI rule

* fix format string for mingw-w64

* attempt to improve/generalize CI rule

* another attempt at fixing format string silliness

* last attempt to fix format string on all archs

+++ Source/net_server.c
@@ -1125,7 +1125,7 @@ static void NET_SV_ParseGameData(net_packet_t *packet, net_client_t *client)
     unsigned int ackseq;
     unsigned int num_tics;
     unsigned int nowtime;
-    size_t i;
+    unsigned long i;
     int player;
     int resend_start, resend_end;
     int index;
